REF: move variety stuff out to typhos.variety
