+++ typhos/widgets.py
@@ -1,3 +1,7 @@
+"""
+Typhos widgets and related utilities.
+"""
+
 import collections
 import datetime
 import inspect
@@ -16,7 +20,8 @@ import pydm.widgets.enum_button
 from ophyd.signal import EpicsSignalBase
 from pydm.widgets.display_format import DisplayFormat
 
-from . import plugins, utils
+from . import plugins, utils, variety
+from .variety import for_variety_read, for_variety_write
 
 logger = logging.getLogger(__name__)
 
@@ -75,65 +80,6 @@ class SignalWidgetInfo(
             write_cls, write_kwargs = widget_type_from_description(obj, desc)
 
         return cls(read_cls, read_kwargs, write_cls, write_kwargs)
-
-
-def _warn_unhandled(instance, metadata_key, value):
-    if value is None:
-        return
-
-    logger.warning(
-        '%s: Not yet implemented variety handling: key=%s value=%s',
-        instance.__class__.__name__, metadata_key, value
-    )
-
-
-def _warn_unhandled_kwargs(instance, kwargs):
-    for key, value in kwargs.items():
-        _warn_unhandled(instance, key, value)
-
-
-def _set_variety_key_handler(key):
-    """
-    A method wrapper to mark a specific variety metadata key with the method.
-
-    Parameters
-    ----------
-    key : str
-        The variety key (e.g., 'delta')
-    """
-
-    def wrapper(method):
-        assert callable(method)
-        if not hasattr(method, '_variety_handler'):
-            method._variety_handler_keys = set()
-        method._variety_handler_keys.add(key)
-        return method
-
-    return wrapper
-
-
-def _get_variety_handlers(members):
-    handlers = {}
-    for attr, method in members:
-        for key in getattr(method, '_variety_handler_keys', []):
-            if key not in handlers:
-                handlers[key] = [method]
-            handlers[key].append(method)
-
-    return handlers
-
-
-def uses_variety_handler(cls):
-    """
-    Class wrapper to finish variety handler configuration.
-
-    Parameters
-    ----------
-    cls : class
-        The class to wrap.
-    """
-    cls._variety_handlers = _get_variety_handlers(inspect.getmembers(cls))
-    return cls
 
 
 class TogglePanel(QWidget):
@@ -198,79 +144,6 @@ class TogglePanel(QWidget):
                 self.contents.show()
             else:
                 self.contents.hide()
-
-
-_variety_to_widget_class = {}
-
-
-def for_variety(variety, *, read=True, write=True):
-    """
-    A class wrapper to associate a specific variety with the class.
-
-    Defaults to registering for both read and write widgets.
-
-    Parameters
-    ----------
-    variety : str
-        The variety (e.g., 'command')
-
-    read : bool, optional
-        Use for readback widgets
-
-    write : bool, optional
-        Use for setpoint widgets
-    """
-
-    known_varieties = {
-        'array-histogram',
-        'array-image',
-        'array-nd',
-        'array-timeseries',
-        'bitmask',
-        'command',
-        'command-enum',
-        'command-proc',
-        'command-setpoint-tracks-readback',
-        'enum',
-        'scalar',
-        'scalar-range',
-        'scalar-tweakable',
-        'text',
-        'text-enum',
-        'text-multiline',
-    }
-
-    if variety not in known_varieties:
-        # NOTE: not kept in sync with pcdsdevices; so this wrapper may need
-        # updating.
-        raise ValueError(f'Not a known variety: {variety}')
-
-    def wrapper(cls):
-        if variety not in _variety_to_widget_class:
-            _variety_to_widget_class[variety] = {}
-
-        if read:
-            _variety_to_widget_class[variety]['read'] = cls
-
-        if write:
-            _variety_to_widget_class[variety]['write'] = cls
-
-        if not read and not write:
-            raise ValueError('`write` or `read` must be set.')
-
-        return cls
-
-    return wrapper
-
-
-def for_variety_read(variety):
-    """`for_variety` shorthand for setting the readback widget class."""
-    return for_variety(variety, read=True, write=False)
-
-
-def for_variety_write(variety):
-    """`for_variety` shorthand for setting the setpoint widget class."""
-    return for_variety(variety, read=False, write=True)
 
 
 @for_variety_write('enum')
@@ -403,7 +276,7 @@ class TyphosLineEdit(pydm.widgets.PyDMLineEdit):
             self.displayFormat = DisplayFormat.Exponential
 
 
-@for_variety('array-nd')
+@for_variety_read('array-nd')
 @for_variety_read('command')
 @for_variety_read('command-enum')
 @for_variety_read('command-proc')
@@ -415,6 +288,7 @@ class TyphosLineEdit(pydm.widgets.PyDMLineEdit):
 @for_variety_read('text')
 @for_variety_read('text-enum')
 @for_variety_read('text-multiline')
+@for_variety_write('array-nd')
 class TyphosLabel(pydm.widgets.PyDMLabel):
     """
     Reimplementation of PyDMLabel to set some custom defaults
@@ -733,7 +607,7 @@ def _create_variety_property():
                     doc='Additional component variety metadata.')
 
 
-@uses_variety_handler
+@variety.uses_variety_handler
 @for_variety_write('scalar-range')
 class TyphosScalarRange(pydm.widgets.PyDMSlider):
     def __init__(self, *args, variety_metadata=None, ophyd_signal=None,
@@ -745,7 +619,7 @@ class TyphosScalarRange(pydm.widgets.PyDMSlider):
 
     variety_metadata = _create_variety_property()
 
-    @_set_variety_key_handler('range')
+    @variety._set_variety_key_handler('range')
     def _variety_key_handler_range(self, value, source, **kwargs):
         """Variety hook for the sub-dictionary "range"."""
         if source == 'value':
@@ -756,23 +630,23 @@ class TyphosScalarRange(pydm.widgets.PyDMSlider):
                 self.userDefinedLimits = True
         # elif source == 'use_limits':
         else:
-            _warn_unhandled(self, 'range.source', source)
+            variety._warn_unhandled(self, 'range.source', source)
 
-        _warn_unhandled_kwargs(self, kwargs)
+        variety._warn_unhandled_kwargs(self, kwargs)
 
-    @_set_variety_key_handler('delta')
+    @variety._set_variety_key_handler('delta')
     def _variety_key_handler_delta(self, value, source, signal=None, **kwargs):
         """Variety hook for the sub-dictionary "delta"."""
         if source == 'value':
             self._delta_value = value
         # elif source == 'signal':
         else:
-            _warn_unhandled(self, 'delta.source', source)
+            variety._warn_unhandled(self, 'delta.source', source)
 
         # range_ = kwargs.pop('range')  # unhandled
-        _warn_unhandled_kwargs(self, kwargs)
+        variety._warn_unhandled_kwargs(self, kwargs)
 
-    @_set_variety_key_handler('display_format')
+    @variety._set_variety_key_handler('display_format')
     def _variety_key_handler_display_format(self, value):
         """Variety hook for the sub-dictionary "delta"."""
         self.displayFormat = getattr(DisplayFormat, value.capitalize(),
@@ -820,22 +694,6 @@ class TyphosScalarRange(pydm.widgets.PyDMSlider):
 class TyphosTweakable(TyphosScalarRange):
     ...
     # TODO tweak functionality from positioner?
-
-
-def _get_widget_class_from_variety(desc, variety_md, read_only):
-    variety = variety_md['variety']  # a required key
-    read_key = 'read' if read_only else 'write'
-    try:
-        widget_cls = _variety_to_widget_class[variety].get(read_key)
-    except KeyError:
-        logger.error('Unsupported variety: %s (%s / %s)', variety,
-                     desc, variety_md)
-    else:
-        if widget_cls is None:
-            # TODO: remove
-            logger.error('TODO no widget?: %s (%s / %s)', variety,
-                         desc, variety_md)
-        return widget_cls
 
 
 def _get_scalar_widget_class(desc, variety_md, read_only):
@@ -899,7 +757,7 @@ def widget_type_from_description(signal, desc, read_only=False):
     }
 
     if variety_metadata:
-        widget_cls = _get_widget_class_from_variety(
+        widget_cls = variety._get_widget_class_from_variety(
             desc, variety_metadata, read_only)
     else:
         try:
